only logged user can delete

+++ src/components/ItemsList.js
@@ -2,7 +2,8 @@ import React, { Component } from "react";
 import { Link, withRouter } from "react-router-dom";
 import axios from "axios";
 import Footer from "./Footer";
-
+import AddItems from "./AddItems";
+import { Form , DropdownButton } from 'react-bootstrap';
 const ClothesItem = (props) => (
   <tr>
     <td>{props.item.itemName}</td>
@@ -10,6 +11,7 @@ const ClothesItem = (props) => (
     <td>{props.item.type}</td>
     <td>{props.item.description}</td>
     <td>{props.item.phoneNumber}</td>
+    <td>{props.item.username}</td>
     <td>
       <img
         src={props.item.image}
@@ -21,16 +23,23 @@ const ClothesItem = (props) => (
     </td>
     <td>
       <Link
+      style={{width:120, height:40}}
         to={"/edit/" + props.item._id}
         className="btn btn-deep-orange darken-4"
       >
         Edit
       </Link>
-      <button
+      <button style={{width:120,height:40}}
         type="button"
         className="btn btn-deep-orange darken-4"
         onClick={() => {
-          props.deleteItem(props.item._id);
+          console.log( props.item, "thiiiis")
+          console.log( window.localStorage.username, " Storaaaage")
+          if(window.localStorage.length > 0 && window.localStorage.username === props.item.username){
+              props.deleteItem(props.item._id);
+          }
+          else {alert("user can only delete the items he added");}
+
         }}
       >
         Delete
@@ -48,9 +57,9 @@ class ItemsList extends Component {
       SearchString: "",
       category: "",
       type: "",
+      username:""
     };
   }
-
   componentDidMount() {
     axios
       .get("http://localhost:3000/addItems/")
@@ -63,19 +72,28 @@ class ItemsList extends Component {
       });
   }
   deleteItem(id) {
-    axios
-      .delete("http://localhost:3000/addItems/" + id)
-      .then((res) => console.log(res.data));
-    this.setState({
-      items: this.state.items.filter((el) => el._id !== id),
-    });
+
+    // console.log( window.localStorage.username, " Storaaaage")
+    // console.log(this.props.items.username, " thisssss")
+  
+          axios
+            .delete("http://localhost:3000/addItems/" + id)
+            .then((res) => console.log(res.data));
+          this.setState({
+            items: this.state.items.filter((el) => el._id !== id),
+          });
+
   }
+  
 
   itemsList() {
+    
     let listedItems =
       this.state.filteredItems.length > 0
         ? this.state.filteredItems
         : this.state.items;
+      
+
     return listedItems.map((currentItem) => {
       return (
         <ClothesItem
@@ -85,8 +103,8 @@ class ItemsList extends Component {
         />
       );
     });
+   
   }
-
   onChangecategory(e) {
     let { items } = this.state;
     let string = e.target.value;
@@ -105,54 +123,45 @@ class ItemsList extends Component {
     filteredItems = filteredItems.filter((item) => item.type.includes(string));
     this.setState({ filteredItems: filteredItems });
   }
-
   render() {
     return (
       <div>
         <br />
         <div className="container text-center border border-light p-9">
           <h2>Clothing</h2>
-          <lable>
-            Select by category..
-            <select
-              ref="userInput"
-              required="true"
-              value={this.state.category}
-              onChange={this.onChangecategory.bind(this)}
-            >
-              <option value=""></option>
-              <option value="Women">Women</option>
-              <option value="Men">Men</option>
-              <option value="Kids">Kids</option>
-            </select>
-          </lable>
-          <lable>
-            Select by type..
-            <select
-              ref="userInput"
-              required="true"
-              value={this.state.type}
-              onChange={this.onChangetype.bind(this)}
-            >
-              <option value=""></option>
-              <option value="Shoes">Shoes</option>
-              <option value="Dress">Dress</option>
-              <option value="Jacket">Jacket</option>
-              <option value="Blouse">Blouse</option>
-              <option value="Gloves">Gloves</option>
-              <option value="Hat">Hat</option>
-              <option value="Scarf">Scarf</option>
-            </select>
-          </lable>
-
+          <Form>
+  <Form.Group controlId="exampleForm.SelectCustomSizeSm"  onChange={this.onChangetype.bind(this)} >
+    {/* <Form.Label>Select by type</Form.Label> */}
+    <Form.Control as="select"  size="sm" custom style={{width:155, color:'white', border:'orange', margin:"50px 0px 10px 250px", background:'#212121'}} >
+    <option value="">Select by type</option>
+    <option value="Shoes">Shoes</option>
+    <option value="Dress">Dress</option>
+    <option value="Jacket">Jacket</option>
+    <option value="Blouse">Blouse</option>
+    <option value="Gloves">Gloves</option>
+    <option value="Hat">Hat</option>
+    <option value="Scarf">Scarf</option>
+    </Form.Control>
+  </Form.Group>
+  <Form.Group controlId="exampleForm.SelectCustomSizeSm" >
+    {/* <Form.Label>Select by category</Form.Label> */}
+    <Form.Control as="select" size="sm" custom onChange={this.onChangecategory.bind(this)} style={{width:155, margin:'-100px 250px 10px 0px', background:'#212121',color:'white'}}>
+    <option value="">Select by category</option>
+    <option value="Women">Women</option>
+    <option value="Men">Men</option>
+    <option value="Kids">Kids</option>
+    </Form.Control>
+  </Form.Group>
+</Form>
           <table className="table">
             <thead className="thead">
               <tr>
-                <th>Item Name</th>
+                <th>Item</th>
                 <th>Category</th>
                 <th>Type</th>
                 <th>Description</th>
                 <th>Donor Phone Number</th>
+                <th>Donor Name</th>
                 <th>Image</th>
               </tr>
             </thead>
@@ -165,3 +174,14 @@ class ItemsList extends Component {
   }
 }
 export default withRouter(ItemsList);
+
+
+
+
+
+
+
+
+
+
+
